ImageMagick builds when HDRI is enabled
git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@17515 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ coders/rle.c
@@ -456,7 +456,7 @@ static Image *ReadRLEImage(const ImageInfo *image_info,ExceptionInfo *exception)
             if (IsValidColormapIndex(image,*p & mask,&index,exception) ==
                 MagickFalse)
               break;
-            *p=colormap[index];
+            *p=colormap[(ssize_t) index];
             p++;
           }
         else
@@ -467,7 +467,7 @@ static Image *ReadRLEImage(const ImageInfo *image_info,ExceptionInfo *exception)
                 if (IsValidColormapIndex(image,(size_t) (x*map_length+
                     (*p & mask)),&index,exception) == MagickFalse)
                   break;
-                *p=colormap[index];
+                *p=colormap[(ssize_t) index];
                 p++;
               }
         if ((i < (ssize_t) number_pixels) || (x < (ssize_t) number_planes))
